requirements.txt
python-dotenv~=1.0.1
langchain-openai~=0.2.1
langchain-anthropic
langchain-community~=0.3.1
chromadb
unstructured[all-docs]~=0.15.13
unstructured-client~=0.25.9
streamlit~=1.38.0
psutil
langchain~=0.3.1
langchain-chroma
langchain-core~=0.3.6
pydantic~=2.9.2